Return earlier if the instance has been deleted
During the reconciliation loop, if an AnsibleEE instance is delete, the
operator keeps trying to recreate it again, failing with several
stacktrace in the attempt, which is pretty normal because the instance
doesn't exist anymore.
This patch let the reconciliation loop returning earlier and gracefully
handling this case.
In addition, the Dockerfile has been updated removing the empty templates/
directory which is not used in this context.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>