Add support for building ee image using local edpm-ansible repo
In order to build ee image from local edpm-ansible changes. It adds
the support for the same by passing the clone directory as a volume
to the build args.

It will be useful in CI jobs where the project is clonned in the
job workspace.

Donot remove the /var/tmp/edpm-ansible directory on local edpm clone.

Signed-off-by: Chandan Kumar <raukadah@gmail.com>