Allow attaching service pods to additional networks
For network isolation outgoing traffic, service pods need to be
attached additional networks. To achive this, the
`k8s.v1.cni.cncf.io/networks` annotation needs to be set on the
pods. This adds a networkAttachments parameter to the
CRD where it can be specified to which additional networks the
service pods should be attached to.
Its a list of pre-created net-attach-defs names. Per default
the pods won't get attached.

```
  ...
  networkAttachments:
  - internalapi
```

Depends-On: openstack-k8s-operators/lib-common#133