Change Task Vars to be []string
When deploying a OpenStackDataPlaneNode, the created openstackansibleee
fails with wrong type for vars:

Identity added: /runner/artifacts/643f357f-5f77-4284-9a48-79df8d67f89c/ssh_key_data (root@osp-converge-01.lab.eng.brq2.redhat.com)
ERROR! Vars in a IncludeRole must be specified as a dictionary, or a list of dictionaries

The error appears to be in '/usr/share/ansible/edpm-playbooks/playbook.yaml': line 13, column 13, but may
be elsewhere in the file depending on the exact syntax problem.

The offending line appears to be:

        tasks_from: main.yml
      vars: ""
            ^ here

This changes the Vars in the Task struct to be of type []string.