allow AEE cr to be created specifying a playbook name
This change fixes two bugs. first the roles paremater is made
optional and converted to a pointer so we can test if its defined.
This enables you to create a AEE cr using the golan types
without specifying a role or inline play.

This also fixes the fact that a job does not have a
redhatcomv1alpha1.AnsibleExecutionJobReadyCondition
at least when its complete.

so we just check if it succeded or failed instead.

Closes: #104