Remove TTLSecondsAfterFinished field
The jobs owned by OpenStackAnsibleEE should not be automatically deleted
by kubernetes. The job's started pods contain the ansible logs, which
are deployment artificats. They should only be deleted when the
OpenStackAnsibleEE itself is deleted, and never automatically.

TTLSecondsAfterFinished must not be set on the Job resource to prevent
automatic deletion. Since the field is no longer needed, it's also
removed from the OpenStackAnsibleEE CRD.

Signed-off-by: James Slagle <jslagle@redhat.com>