Cleanup golangci-lint
Since the controller runtime and golang bump the golangci-lint github
action started to fail. However golangci-lint also run in pre-commit
both locally and in CI. So the github action is removed. Also I took the
opportunity to bump the golangci-lint version to the newest that still
supports 1.20. And removed disabled checks that passing now.