Make AEE containers resources configurable
"Creating a pod without CPU requests or limits will effectively allow it to be scheduled on any suitable node, regardless of the amount of CPU left on that node. In practice, it will still get some minimal CPU guarantees. However, it will not be able to use much CPU time when there is node pressure"

https://www.datadoghq.com/blog/kubernetes-cpu-requests-limits/
Signed-off-by: Fabricio Aguiar <fabricio.aguiar@gmail.com>