Don't require in-line Play text to start with list
This change removes the webhook validation requirement that in-line plays
need to be lists. A play in Ansible is a dict, while a playbook is a list of plays. This change
ensures we validate on the play format to conform with the terminology in use. For reference, see:
https://github.com/ansible/ansible/blob/10f9b8e6554e024e3561170153b8e7fde5e7e4fb/test/units/playbook/test_play.py#L48

Given the terminology we use in the OpenStackDataPlaneService of 'play', we should then validate
on the Ansible defined 'play' format, rather than a 'playbook' format.

Signed-off-by: Brendan Shephard <bshephar@redhat.com>